requirements.txt
pytesseract
opencv-python-headless
opencv-python
pytesseract
pillow
